+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.2.14] - 2025-04-30
+- Support new-line separated headers for Rack 2 backward compatibility.
+
+## [0.2.12] - 2025-04-29
+- Max Rust edition is now "2021"
+- Removed invalid rbs files causing RI doc generation failure
+- Fixed header clobbering in Rack
+- Added new `ruby_thread_request_backlog_size` option
+
 ## [0.2.3] - 2025-04-22
 
 - Public release!